refactor background wrapper and and extract icons into components
