+++ react_native_frontend/src/components/form/index.js
@@ -1,5 +1,7 @@
-import Button from './Button';
-import Input from './Input';
+import Button from './button';
+import Input from './input';
+import BackButton from './backButton';
 
-export { Button, Input };
-export default { Button, Input };
+
+export { Button, Input, BackButton };
+export default { Button, Input, BackButton };
