+++ react_native_frontend/src/components/index.js
@@ -1,10 +1,10 @@
 import Form, { Input, Button } from './form';
-import Partials, { AlertStatus, BackgroundWrapper, Logo } from './partials';
+import Partials, { AlertStatus, BackButton, Logo } from './partials';
 import Text, { TextFont, Heading } from './text';
 
 export {
     Form, Input, Button,
-    Partials, AlertStatus, BackgroundWrapper, Logo,
+    Partials, AlertStatus, BackButton, Logo,
     Text, TextFont, Heading
 };
 
@@ -14,7 +14,7 @@ export default {
     Button,
     Partials,
     AlertStatus,
-    BackgroundWrapper,
+    BackButton,
     Logo,
     Text,
     TextFont,
