+++ react_native_frontend/src/components/partials/alertStatus.js
