+++ react_native_frontend/src/components/partials/logo.js
