+++ react_native_frontend/src/components/text/heading.js
