+++ react_native_frontend/src/pages/index.js
@@ -1,6 +1,6 @@
-import Home from './Home';
-import Login from './Login';
-import Register from './Register';
+import Home from './home';
+import Login from './login';
+import Register from './register';
 
 export { Home, Login, Register };
 export default { Home, Login, Register };
